Move Landing component to Landing/index.jsx
